Restructured the working files to be up to date to all commits and merges made into the remote group repo
moved darius work from the user controller into the other boilerplate made by uriah.

uriah's work is still in the boilerplate from the models.
