+++ Pipfile
@@ -5,8 +5,10 @@ name = "pypi"
 
 [packages]
 flask = "*"
+pymysql = "*"
+flask-bcrypt = "*"
 
 [dev-packages]
 
 [requires]
-python_version = "3.8"
+python_version = "3.12"
