+++ flask_app/__init__.py
@@ -1,4 +1,6 @@
 from flask import Flask
-
 app = Flask(__name__)
-app.secret_key = "9f0f95663a2b1a24336e40279d9332727c8c9db9cba6c4fc3a13159ce66e961d"
+app.secret_key = "The first rule of secret key is..." 
+
+# The secret key is needed to run session
+# This is one thing that would usually be kept in your git ignore, along with API keys
