--- hello_flask/habit_tracker/README.md
@@ -1,2 +1,0 @@
-# Habit-Tracker-Group-Project
-Competition based habit tracker. Users will choose between several different habit categories and in the future receive rewards based on the length of their tracking streaks.
