--- hello_flask/habit_tracker/flask_app/instructions.txt
@@ -1,5 +1,0 @@
-Check the pipfile to see if it matches your python version.
-Change the version in the pipfile if it doesn't match. 
-In the terminal from the directory level of server.py run"
-python3 -m pipenv install
-or your computer's equavilant
